Moved some things around to make structure file shorter
1. Created State class and moved refactored collapseState to Collapse.instance(State)
2. Moved Collapse.instance(Tree to Tree file
